+++ src/main/java/org/noostak/group/application/create/GroupCreateServiceImpl.java
@@ -1,6 +1,7 @@
-package org.noostak.group.application;
+package org.noostak.group.application.create;
 
 import lombok.RequiredArgsConstructor;
+import org.noostak.group.application.InvitationCodeGenerator;
 import org.noostak.group.common.exception.GroupErrorCode;
 import org.noostak.group.common.exception.GroupException;
 import org.noostak.group.domain.GroupRepository;
